# apispec/api.yaml
openapi: 3.0.4 # TODO: Update to 3.1.x
info:
  title: Leaderboard Apex Service API
  version: 1.0.0
  description: An API for storing and retrieving player scores.
...
paths:
...
  /api/v1/boards/{board_id}/scores/{player_id}:
    get:
      tags: [Scores]
      summary: Retrieves a player's standing on a board (rank, score, total)
      parameters:
        - $ref: "#/components/parameters/BoardIdPath"
        - $ref: "#/components/parameters/PlayerIdPath"
      responses:
        "200":
          description: All ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: ID not found (player not ranked on this board)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags: [Scores]
      summary: Sets a player's absolute score
      description: >
        The score is a integer number within +-1e13 (ten trillion).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetScoreRequest"
      parameters:
        - $ref: "#/components/parameters/BoardIdPath"
        - $ref: "#/components/parameters/PlayerIdPath"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "204":
          description: Score updated (no content)
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: ID not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Board is closed, or the Idempotency-Key was reused with a different request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          description: >-
            What went wrong.
...
    SetScoreRequest:
      type: object
      additionalProperties: false
      required:
        - player_score
      properties:
        player_score:
          type: integer
          format: int64
          minimum: -10000000000000
          maximum: 10000000000000
          description: >
            Absolute score. Integer number within +-1e13 (ten trillion):
            other values will be rejected. Note that unix second and millisecond timestamps fit.
...
    TotalMetadata:
      type: object
      additionalProperties: false
      required:
        - total
      properties:
        total:
          type: integer
          description: Total ranked players on the board
    ScoreEntry:
      type: object
      additionalProperties: false
      required:
        - player_id
        - score
        - rank
      properties:
        player_id:
          type: string
        score:
          type: integer
          format: int64
          minimum: -10000000000000
          maximum: 10000000000000
        rank:
          type: integer
          description: 1-based position on the leaderboard (rank 1 = highest score)
    RankResponse:
      type: object
      additionalProperties: false
      required:
        - standing
        - metadata
      properties:
        standing:
          $ref: "#/components/schemas/ScoreEntry"
        metadata:
          $ref: "#/components/schemas/TotalMetadata"
...
  parameters:
    PlayerIdPath:
      name: player_id
      in: path
      required: true
      description: player ID
      schema:
        type: string
    BoardIdPath:
      name: board_id
      in: path
      required: true
      description: leaderboard ID
      schema:
        type: string
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Optional key that makes a retried write a no-op. Reusing a key with a
        different operation or amount is rejected with 409. Records expire after 24h.
      schema:
        type: string
        maxLength: 128
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PayloadTooLarge:
      description: Request body exceeds the size limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"